+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ## 4.24.0 (Unreleased)
 
 ### Added
-- Support for opting out of DNS when launching `core_instance` added
+- Support for opting out of DNS when launching `core_instance` 
 - Support for Cluster Network Resize
 - Support for Intelligent Tiering in added to `Object Storage`
 - Support for Relax Parameters for VLAN and make vcnId as optional for LIST VLAN API
 - Support for CloudGuard features release
 - Support for RAC External Database for DB Management
+- Support for Application Performance Monitoring for `functions` 
 
 ## 4.23.0 (April 21, 2021)